--- src/App.css
@@ -1,167 +1,0 @@
-/* http://meyerweb.com/eric/tools/css/reset/ 
-   v2.0 | 20110126
-   License: none (public domain)
-*/
-
-html, body, div, span, applet, object, iframe,
-h1, h2, h3, h4, h5, h6, p, blockquote, pre,
-a, abbr, acronym, address, big, cite, code,
-del, dfn, em, img, ins, kbd, q, s, samp,
-small, strike, strong, sub, sup, tt, var,
-b, u, i, center,
-dl, dt, dd, ol, ul, li,
-fieldset, form, label, legend,
-table, caption, tbody, tfoot, thead, tr, th, td,
-article, aside, canvas, details, embed, 
-figure, figcaption, footer, header, hgroup, 
-menu, nav, output, ruby, section, summary,
-time, mark, audio, video {
-	margin: 0;
-	padding: 0;
-	border: 0;
-	font-size: 100%;
-	font: inherit;
-	vertical-align: baseline;
-}
-/* HTML5 display-role reset for older browsers */
-article, aside, details, figcaption, figure, 
-footer, header, hgroup, menu, nav, section {
-	display: block;
-}
-body {
-	line-height: 1;
-}
-ol, ul {
-	list-style: none;
-}
-blockquote, q {
-	quotes: none;
-}
-blockquote:before, blockquote:after,
-q:before, q:after {
-	content: '';
-	content: none;
-}
-table {
-	border-collapse: collapse;
-	border-spacing: 0;
-}
-
-@font-face {
-  font-family: 'PokeTitle';
-  src: local('PokeTitle'), url(./fonts/Pocket-Monk.otf);
-}
-@font-face {
-  font-family: 'FireRed';
-  src: local('FireRed'), url(./fonts/pokemon_fire_red.otf);
-}
-
-.mainView {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.headerSpacer {
-  background-color: #CC0000;
-  height: 11vh;
-}
-
-.searchBar {
-  margin-top: 6vh;
-  border: 2px solid var(--details);
-  border-radius: 50px;
-  height: 40px;
-  width: 400px;
-  padding: 2px;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-bottom: 6vh;
-}
-
-.container {
-  height: 100%;
-  width: 100%;
-  vertical-align: middle;
-}
-
-.searchBox {
-  background-color: var(--bg);
-  border: none;
-  height: 100%;
-  width: 100%;
-  padding: 0px 15px 4.5px;
-  border-radius: 50px;
-  font-size: 33px;
-  font-family: "FireRed", sans-serif;
-  color: var(--details);
-}
-
-.searchBox:focus {
-  outline: none;
-}
-
-.searchIcon {
-  color: var(--details);
-  height: 80%;
-  width: 80%;
-  margin-top: 2.5px;
-  margin-right: 15px; 
-  margin-left: 10px; 
-}
-
-.searchIcon:hover {
-  cursor: pointer;
-}
-
-.searchResults {
-  background-color: var(--bg);
-  border-radius: 20px;
-  padding: 10px;
-  box-shadow: 0px 8px 10px rgba(29, 29, 29, 0.699);
-}
-
-.pokeData {
-  font-family: 'FireRed', sans-serif;
-  font-size: 25px;
-  color: var(--vibrant);
-}
-
-.divider {
-  background-color: var(--details);
-  margin-right: 8px;
-  margin-left: 8px;
-  width: 1px;
-  height: 80px;
-}
-
-.noResult {
-  padding: 0px 20px 0px 20px;
-  color: var(--details);
-  font-family: 'FireRed', sans-serif;
-  font-size: 25px;
-}
-
-header {
-  top: 0;
-  left: 0;
-  position: absolute;
-  box-shadow: 0px 11px #CC0000;
-  width: 100%;
-  background-color: #FF0000;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-h1 {
-  font-family: 'PokeTitle';
-  font-size: 8vh;
-  color: #FFDE00;
-  -webkit-text-stroke-width: 1.3px;
-  -webkit-text-stroke-color: #3B4CCA;
-  text-shadow: -4px 3px 1px #273286;
-  margin: 0;
-  padding: 10px;
-}
