+++ src/components/search/SearchPokemon.css
@@ -1,0 +1,32 @@
+
+
+@font-face {
+  font-family: 'PokeTitle';
+  src: local('PokeTitle'), url(../../assets/fonts/Pocket-Monk.otf);
+}
+@font-face {
+  font-family: 'FireRed';
+  src: local('FireRed'), url(../../assets/fonts/pokemon_fire_red.otf);
+}
+  
+.divider {
+    background-color: var(--details);
+    margin-right: 8px;
+    margin-left: 8px;
+    width: 1px;
+    height: 80px;
+  }
+
+  .pokeData {
+    font-family: 'FireRed', sans-serif;
+    font-size: 25px;
+    color: var(--vibrant);
+  }
+
+  .noResult {
+    padding: 0px 20px 0px 20px;
+    color: var(--details);
+    font-family: 'FireRed', sans-serif;
+    font-size: 25px;
+  }
+
