+++ src/containers/App/App.css
@@ -1,0 +1,44 @@
+
+
+@font-face {
+  font-family: 'PokeTitle';
+  src: local('PokeTitle'), url(../../assets/fonts/Pocket-Monk.otf);
+}
+@font-face {
+  font-family: 'FireRed';
+  src: local('FireRed'), url(../../assets/fonts/pokemon_fire_red.otf);
+}
+
+.mainView {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.headerSpacer {
+  background-color: #CC0000;
+  height: 11vh;
+}
+
+header {
+  top: 0;
+  left: 0;
+  position: absolute;
+  box-shadow: 0px 11px #CC0000;
+  width: 100%;
+  background-color: #FF0000;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+h1 {
+  font-family: 'PokeTitle';
+  font-size: 8vh;
+  color: #FFDE00;
+  -webkit-text-stroke-width: 1.3px;
+  -webkit-text-stroke-color: #3B4CCA;
+  text-shadow: -4px 3px 1px #273286;
+  margin: 0;
+  padding: 10px;
+}
